chore: Moved app related to src and updated imports
