+++ src/app/(dashboard)/analytics/page.tsx
