+++ src/app/(dashboard)/settings/page.tsx
