+++ src/stores/board-store.ts
